logindemo/requirements.txt
asgiref==3.5.2
Django==3.2.12
django-phonenumber-field==5.0.0
phonenumberslite==8.12.46
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
